Fix conversion warnings in FPS calculation
Added explicit static_cast<float>() conversions to eliminate compiler
warnings about implicit conversion from long long and int to float.
This makes the type conversions intentional and clear.

+++ src/entities/FPSEntity.cpp
@@ -56,8 +56,8 @@ void FPSEntity::update(Time deltaTime) {
         return;
 
     // Calculate actual FPS based on real elapsed time
-    auto elapsedSeconds = elapsed.count() / 1000.0f;
-    auto fps = this->framesPassed / elapsedSeconds;
+    auto elapsedSeconds = static_cast<float>(elapsed.count()) / 1000.0f;
+    auto fps = static_cast<float>(this->framesPassed) / elapsedSeconds;
 
     this->statsText->setString("FPS: " + to_string((int)fps) + "\n");
